Add module for notification

+++ be/app/Repositories/CommonAuthRepository.php
@@ -15,6 +15,7 @@ use App\Mail\ForgetPassword;
 use App\Models\ForgetPasswordOtp;
 use Carbon\Carbon;
 use Exception;
+use Modules\Notification\Entities\Notification;
 
 class CommonAuthRepository extends BaseRepository
 {
@@ -136,14 +137,22 @@ class CommonAuthRepository extends BaseRepository
                         'otp' => $uniqueCode,
                         'model_id' => $user->id,
                         'model_type'  => $user::class,
-                        'otp_expire_date_time' => $expireTime
+                        'otp_expire_date_time' => $expireTime,
                     ];
 
                     $otpData = ForgetPasswordOtp::create($data);
                 }
+                $mailTemplate = view('template.forget-password',compact('otpData','user'))->render();
+                
+                $sendMail = [
+                    'description' => $mailTemplate,
+                    'title' => 'Forget Password',
+                    'user' => $user,
+                    'send_by' => 1
+                ];
 
                 // Send email to user
-                Mail::to($user->email)->send(new ForgetPassword($otpData));
+                Notification::createNotification($sendMail);
 
                 $data = [
                     'email' => $user->email
@@ -177,31 +186,41 @@ class CommonAuthRepository extends BaseRepository
             $user = $this->model->where(['email' => $request->email])->first();
             if ($user) {
                 $userOtp = $user->hasOtp;
-                $checkOtpTime = $userOtp->otp_expire_date_time;
-                $currentTime = Carbon::now();
-                if ($currentTime > $checkOtpTime) {
-                    return array('message' => 'Otp have expired', 'statusCode' => 200, 'data' => []);
-                } else {
-                    if ($request->otp == $userOtp->otp) {
-                        $userOtp->update(['is_verify' => 1, 'verify_at' => Carbon::now()]);
-                        return [
-                            'message' => 'Otp verified succesfully',
-                            'statusCode' => 200,
-                            'data' =>     [
-                                'is_verified_otp' => true,
-                                'email' => $user->email
-                            ]
-                        ];
+
+                if ( !empty($userOtp) ) {
+
+                    $checkOtpTime = $userOtp->otp_expire_date_time;
+                    $currentTime = Carbon::now();
+                    if ($currentTime > $checkOtpTime) {
+                        return array('message' => 'Otp have expired', 'statusCode' => 200, 'data' => []);
                     } else {
-                        return [
-                            'message' => 'You have entered wrong otp',
-                            'statusCode' => 200,
-                            'data' =>     [
-                                'is_verified_otp' => false,
-                                'email' => $user->email
-                            ]
-                        ];
+                        if ($request->otp == $userOtp->otp) {
+                            $userOtp->update(['is_verify' => 1, 'verify_at' => Carbon::now()]);
+                            return [
+                                'message' => 'Otp verified succesfully',
+                                'statusCode' => 200,
+                                'data' =>     [
+                                    'is_verified_otp' => true,
+                                    'email' => $user->email
+                                ]
+                            ];
+                        } else {
+                            return [
+                                'message' => 'You have entered wrong otp',
+                                'statusCode' => 200,
+                                'data' =>     [
+                                    'is_verified_otp' => false,
+                                    'email' => $user->email
+                                ]
+                            ];
+                        }
                     }
+                } else {
+                    return [
+                        'message' => 'OTP not generated yet',
+                        'statusCode' => 200,
+                        'data' =>     []
+                    ];
                 }
             }
         } catch (Exception $e) {
@@ -218,25 +237,43 @@ class CommonAuthRepository extends BaseRepository
 
             $user = $this->model->where(['email' => $request->email])->first();
             if ($user) {
+
                 $userOtp = $user->hasOtp;
-                $verify_time_expire = (new Carbon($userOtp->verify_at))->addMinutes(5); // If you have verfiy your otp but do not forget the password so In forget password screen is valid for next 5 minutes after that it give expire error
-                $current_time = Carbon::now();
-                if ( !empty($userOtp) && !empty($userOtp->is_verify) && $current_time < $verify_time_expire) {
 
-                    $password = Hash::make($request->password);
+                if ( $userOtp ) {
 
-                    $user->update(['password' => $password]);
-                    $userOtp->delete();
-                    return [
-                        'message' => 'Password Change Successfully',
-                        'statusCode' => 200,
-                    ];
+                    if ( !$userOtp->is_verify ) {
+                        return [
+                            'message' => 'verify the otp first',
+                            'statusCode' => 200,
+                        ];
+                    }
+
+                    $verify_time_expire = (new Carbon($userOtp->verify_at))->addMinutes(5); // If you have verfiy your otp but do not forget the password so In forget password screen is valid for next 5 minutes after that it give expire error
+                    $current_time = Carbon::now();
+                    if ( !empty($userOtp) && !empty($userOtp->is_verify) && $current_time < $verify_time_expire) {
+
+                        $password = Hash::make($request->password);
+
+                        $user->update(['password' => $password]);
+                        $userOtp->delete();
+                        return [
+                            'message' => 'Password Change Successfully',
+                            'statusCode' => 200,
+                        ];
+                    } else {
+                        return [
+                            'message' => 'Your otp verfication time have been expired .please again verify otp',
+                            'statusCode' => 200
+                        ];
+                    }
                 } else {
                     return [
-                        'message' => 'Your otp verfication time have been expired .please again verify otp',
+                        'message' => 'OTP not generated yet',
                         'statusCode' => 200
                     ];
                 }
+
             } else {
 
                 return [
